refactor(cli): extract suppression subcommand into cli/commands/suppression.mjs
Move the `suppression` (suppression add) block out of main() into a dedicated
handler. Behavior, messages, and exit codes are unchanged; only the enclosing
function and relative import depth differ.
